@@ -1,16 +1,13 @@
-/* global describe, it, xit, before, beforeEach, after, afterEach */
 /* jslint node: true, esnext: true */
 
 'use strict';
 
-const assert = require('chai').assert;
-
+import test from 'ava';
 import Context from '../src/Context';
 import MergeLineSet from '../src/MergeLineSet';
-
 import Client from './Client';
 
-describe('merge lines', () => {
+test('merge lines', async t => {
   const context = new Context(new Client({
     'aFile': {
       templateRepo: ['Line 1', 'Line 2'].join('\n'),
@@ -19,8 +16,6 @@ describe('merge lines', () => {
   }), 'targetRepo', 'templateRepo', {});
 
   const merger = new MergeLineSet(context, 'aFile');
-
-  it('lines', () =>
-    merger.merge.then(m =>
-      assert.equal(m.content, ['Line 1', 'Line 2', 'Line 3'].join('\n'))));
+  const merged = await merger.merge;
+  t.deepEqual(merged.content, ['Line 1', 'Line 2', 'Line 3'].join('\n'));
 });
