@@ -1,16 +1,14 @@
-/* global describe, it, xit, before, beforeEach, after, afterEach */
 /* jslint node: true, esnext: true */
 
 'use strict';
 
-const assert = require('chai').assert;
-
+import test from 'ava';
 import Context from '../src/Context';
 import Readme from '../src/Readme';
 import Package from '../src/Package';
 import Client from './Client';
 
-describe('readme', () => {
+test('readme', async t => {
   const context = new Context(new Client({
     'aFile': {
       templateRepo: `Line 1x
@@ -28,9 +26,8 @@ Line 2`
   context.addFile(new Package(context, 'package.json'));
 
   const readme = new Readme(context, 'aFile');
+  const merged = await readme.merge;
 
-  it('lines', () =>
-    readme.merge.then(m =>
-      assert.equal(m.content, `
-Line 2`)));
+  t.deepEqual(merged.content, `
+Line 2`);
 });
