@@ -1,16 +1,13 @@
-/* global describe, it, xit, before, beforeEach, after, afterEach */
 /* jslint node: true, esnext: true */
 
 'use strict';
 
-const assert = require('chai').assert;
-
+import test from 'ava';
 import Context from '../src/Context';
 import Replace from '../src/Replace';
-
 import Client from './Client';
 
-describe('replace', () => {
+test('replace', async t => {
   const context = new Context(new Client({
     'aFile': {
       templateRepo: `Line 1x
@@ -21,9 +18,7 @@ describe('replace', () => {
   }), 'targetRepo', 'templateRepo', {});
 
   const replace = new Replace(context, 'aFile');
-
-  it('lines', () =>
-    replace.merge.then(m =>
-      assert.equal(m.content, `Line 1x
-        Line 2x`)));
+  const merged = await replace.merge;
+  t.deepEqual(merged.content, `Line 1x
+        Line 2x`);
 });
